+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 ### Unreleased
-- Reset configuration before new configuration 
+- Reset configuration before new configuration
+- Encrypt initiator password with X506 Certificate key